+++ ving/tests/session.api.test.mjs
@@ -1,4 +1,4 @@
-import axios from 'axios';
+import { ofetch } from "ofetch";
 import { useKind } from '#ving/record/utils.mjs';
 import { describe, test, expect, afterAll } from "vitest";
 import { like, eq, asc, desc, and } from '#ving/drizzle/orm.mjs';
@@ -10,38 +10,44 @@ const base = `http://localhost:3000/api/${vingConfig.rest.version}/`;
 
 describe('Session API', async () => {
     await Users.delete.where(eq(Users.table.username, 'rita'));
-    const user = (await axios.post(
+    const user = (await ofetch(
         base + 'user?includeMeta=true',
-        { username: 'rita', realName: 'Rita Hayworth', email: 'rita@shawshank.jail', password: 'poster' },
-    )).data;
+        {
+            method: "POST",
+            body: { username: 'rita', realName: 'Rita Hayworth', email: 'rita@shawshank.jail', password: 'poster' },
+        }
+    ));
 
     test('create user', () => {
         expect(user.meta.displayName).toBe('rita');
     });
 
-    const session = (await axios.post(
+    const session = (await ofetch(
         base + 'session',
-        { login: 'rita', password: 'poster', sessionType: 'native' }
-    )).data;
+        {
+            method: "POST",
+            body: { login: 'rita', password: 'poster', sessionType: 'native' },
+        }
+    ));
 
     test('login', () => {
         expect(session.props.userId).toBe(user.props.id);
     });
 
     test('get session', async () => {
-        const result = (await axios.get(
+        const result = (await ofetch(
             `${base}session/${session.props.id}?includeRelated=user`,
             { headers: { Cookie: `vingSessionId=${session.props.id}` } }
-        )).data;
+        ));
         expect(result.related.user.props.username).toBe('rita');
 
     });
 
     test('logout', async () => {
-        const result = (await axios.delete(
+        const result = (await ofetch(
             `${base}session`,
-            { headers: { Cookie: `vingSessionId=${session.props.id}` } }
-        )).data;
+            { method: 'DELETE', headers: { Cookie: `vingSessionId=${session.props.id}` } }
+        ));
         expect(result.props.id).toBe(session.props.id);
     });