+++ ving/tests/test.api.test.mjs
@@ -1,5 +1,5 @@
 import { describe, test, expect } from "vitest";
-import axios from 'axios'
+import { ofetch } from "ofetch";
 import { getConfig } from '#ving/config.mjs';
 
 const vingConfig = await getConfig();
@@ -7,8 +7,7 @@ const base = `http://localhost:3000/api/${vingConfig.rest.version}/`;
 describe('Test API', async () => {
 
     test(`get ${base}`, async () => {
-        const result = await axios(base + 'test')
-        expect(result.status).toBe(200);
-        expect(result.data.success).toBe(true);
+        const result = await ofetch(base + 'test')
+        expect(result.success).toBe(true);
     });
 })